\name{createPDFfromPS}
\alias{createPDFfromPS}
\title{Create a pdf from a set of postscript files.}
\usage{
createPDFfromPS(pdfFile, psFiles, shFile = "runPStoPDF.sh")
}
\arguments{
  \item{pdfFile}{- path to output pdf file}

  \item{psFiles}{- vector of paths to the input postscript
  files}
}
\description{
Function to create a pdf file from a set of postscript
files.
}
\details{


??
}

